+++ SourcesAXR/xrGameSoC/helicopter.h
@@ -23,7 +23,7 @@ enum EHeliHuntState{eEnemyNone,eEnemyPoint,eEnemyEntity};
 struct SHeliEnemy{
 	EHeliHuntState					type;
 	Fvector							destEnemyPos;
-	u32								destEnemyID;
+	u16								destEnemyID;
 	float							fire_trail_length_curr;
 	float							fire_trail_length_des;
 	bool							bUseFireTrail;
